backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Magnesium- <yyong007@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.114.0"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.34"
psycopg2 = "^2.9.9"
prometheus-client = "^0.20.0"
python-multipart = "^0.0.9"
pyjwt = "^2.9.0"
pydantic = {extras = ["email"], version = "^2.9.1"}
pydantic-settings = "^2.4.0"
user-agents = "^2.2.0"
requests = "^2.32.3"
chardet = "^5.2.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.4"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
